[Frontend] Add oodikone-shared
