+++ docker-compose.yml
@@ -60,6 +60,7 @@ services:
     volumes:
       - frontend-nodemod:/opt/app-root/src/node_modules
       - ./services/frontend/:/opt/app-root/src
+      - ./services/oodikone-shared/:/opt/app-root/oodikone-shared
 
   kone-db:
     <<: *defaults
