+++ services/frontend/Dockerfile
@@ -29,9 +29,10 @@ ENV REACT_APP_SENTRY_DSN=$SENTRY_DSN
 ARG SERVICE_PROVIDER
 ENV REACT_APP_SERVICE_PROVIDER=$SERVICE_PROVIDER
 
-COPY ./package* ./
+COPY ./oodikone-shared/ ../oodikone-shared
+COPY ./frontend/package* ./
 RUN npm ci --include=dev
-COPY . .
+COPY ./frontend .
 
 RUN npm run build && npm install -g serve@12.0.0
