+++ services/frontend/package.json
@@ -15,6 +15,7 @@
     "immer": "^10.0.4",
     "lodash": "^4.17.21",
     "moment": "^2.30.1",
+    "oodikone-shared": "file:../oodikone-shared",
     "prop-types": "^15.8.1",
     "query-string": "^9.0.0",
     "react": "^18.2.0",
