+++ services/frontend/vite.config.ts
@@ -23,6 +23,7 @@ export default defineConfig({
     alias: {
       '@': path.resolve(__dirname, './src'),
     },
+    preserveSymlinks: true,
   },
   build: {
     outDir: 'build',
