+++ services/oodikone-shared/index.js
@@ -1,0 +1,6 @@
+/* eslint-disable no-console */
+const helloWorld = () => {
+  console.log('Hello World!')
+}
+
+module.exports = { helloWorld }
